redis-nosql-ingestion/requirements.txt
redis==4.5.5
python-dotenv==1.0.0
python-dateutil==2.8.2